+++ core/services/bulletprooftxmanager/eth_broadcaster.go
@@ -3,6 +3,7 @@ package bulletprooftxmanager
 import (
 	"context"
 	"fmt"
+	"math/big"
 	"sync"
 	"time"
 
@@ -14,10 +15,10 @@ import (
 	"github.com/smartcontractkit/chainlink/core/store/orm"
 	"github.com/smartcontractkit/chainlink/core/utils"
 
+	ethereum "github.com/ethereum/go-ethereum"
 	gethCommon "github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 )
 
 // EthBroadcaster monitors eth_txes for transactions that need to
@@ -214,7 +215,7 @@ func (eb *ethBroadcaster) processUnstartedEthTxs(fromAddress gethCommon.Address)
 			return nil
 		}
 		n++
-		a, err := newAttempt(eb.store, *etx, eb.config.EthGasPriceDefault())
+		a, err := newAttempt(eb.store, *etx, eb.initialTxGasPrice())
 		if err != nil {
 			return errors.Wrap(err, "processUnstartedEthTxs failed")
 		}
@@ -227,6 +228,13 @@ func (eb *ethBroadcaster) processUnstartedEthTxs(fromAddress gethCommon.Address)
 			return errors.Wrap(err, "processUnstartedEthTxs failed")
 		}
 	}
+}
+
+func (eb *ethBroadcaster) initialTxGasPrice() *big.Int {
+	if eb.config.OptimismGasFees() {
+		return big.NewInt(1)
+	}
+	return eb.config.EthGasPriceDefault()
 }
 
 // handleInProgressEthTx checks if there is any transaction
@@ -266,6 +274,16 @@ func getInProgressEthTx(store *store.Store, fromAddress gethCommon.Address) (*mo
 func (eb *ethBroadcaster) handleInProgressEthTx(etx models.EthTx, attempt models.EthTxAttempt, initialBroadcastAt time.Time) error {
 	if etx.State != models.EthTxInProgress {
 		return errors.Errorf("invariant violation: expected transaction %v to be in_progress, it was %s", etx.ID, etx.State)
+	}
+
+	if eb.config.OptimismGasFees() {
+		callMsg := ethereum.CallMsg{To: &etx.ToAddress, From: etx.FromAddress, Data: etx.EncodedPayload}
+		gasLimit, err := eb.ethClient.EstimateGas(context.TODO(), callMsg)
+		if err != nil {
+			return err
+		}
+		etx.GasLimit = gasLimit
+		attempt.GasPrice = *utils.NewBig(big.NewInt(1))
 	}
 
 	sendError := sendTransaction(context.TODO(), eb.ethClient, attempt)
@@ -391,10 +409,10 @@ func (eb *ethBroadcaster) saveInProgressTransaction(etx *models.EthTx, attempt *
 	}
 	etx.State = models.EthTxInProgress
 	return eb.store.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Omit(clause.Associations).Create(attempt).Error; err != nil {
+		if err := tx.Create(attempt).Error; err != nil {
 			return errors.Wrap(err, "saveInProgressTransaction failed to create eth_tx_attempt")
 		}
-		return errors.Wrap(tx.Omit(clause.Associations).Save(etx).Error, "saveInProgressTransaction failed to save eth_tx")
+		return errors.Wrap(tx.Save(etx).Error, "saveInProgressTransaction failed to save eth_tx")
 	})
 }
 
@@ -423,10 +441,10 @@ func saveAttempt(store *store.Store, etx *models.EthTx, attempt models.EthTxAtte
 		if err := IncrementNextNonce(tx, etx.FromAddress, *etx.Nonce); err != nil {
 			return errors.Wrap(err, "saveUnconfirmed failed")
 		}
-		if err := tx.Omit(clause.Associations).Save(etx).Error; err != nil {
+		if err := tx.Save(etx).Error; err != nil {
 			return errors.Wrap(err, "saveUnconfirmed failed to save eth_tx")
 		}
-		if err := tx.Omit(clause.Associations).Save(&attempt).Error; err != nil {
+		if err := tx.Save(&attempt).Error; err != nil {
 			return errors.Wrap(err, "saveUnconfirmed failed to save eth_tx_attempt")
 		}
 		for _, f := range callbacks {
@@ -474,7 +492,7 @@ func saveFatallyErroredTransaction(store *store.Store, etx *models.EthTx) error 
 		if err := tx.Exec(`DELETE FROM eth_tx_attempts WHERE eth_tx_id = ?`, etx.ID).Error; err != nil {
 			return errors.Wrapf(err, "saveFatallyErroredTransaction failed to delete eth_tx_attempt with eth_tx.ID %v", etx.ID)
 		}
-		return errors.Wrap(tx.Omit(clause.Associations).Save(etx).Error, "saveFatallyErroredTransaction failed to save eth_tx")
+		return errors.Wrap(tx.Save(etx).Error, "saveFatallyErroredTransaction failed to save eth_tx")
 	})
 }