+++ core/services/keeper/registry_synchronizer_process_logs.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"fmt"
 	"sync"
 
 	"github.com/pkg/errors"
@@ -43,9 +44,6 @@ func (rs *RegistrySynchronizer) handleSyncRegistryLog(head models.Head, done fun
 	if was {
 		return
 	}
-	if !head.IsInChain(broadcast.RawLog().BlockHash) {
-		return
-	}
 	_, err = rs.syncRegistry()
 	if err != nil {
 		logger.Error(errors.Wrapf(err, "RegistrySynchronizer: unable to sync registry, jobID: %d", rs.job.ID))
@@ -78,9 +76,6 @@ func (rs *RegistrySynchronizer) handleUpkeepCanceledLogs(head models.Head, done 
 		if was {
 			continue
 		}
-		if !head.IsInChain(broadcast.RawLog().BlockHash) {
-			continue
-		}
 		log, ok := broadcast.DecodedLog().(*keeper_registry_wrapper.KeeperRegistryUpkeepCanceled)
 		if !ok {
 			logger.Errorf("RegistrySynchronizer: invariant violation, expected UpkeepCanceled log but got %T", log)
@@ -88,11 +83,12 @@ func (rs *RegistrySynchronizer) handleUpkeepCanceledLogs(head models.Head, done 
 		}
 		ctx, cancel := postgres.DefaultQueryCtx()
 		defer cancel()
-		err = rs.orm.BatchDeleteUpkeepsForJob(ctx, rs.job.ID, []int64{log.Id.Int64()})
+		affected, err := rs.orm.BatchDeleteUpkeepsForJob(ctx, rs.job.ID, []int64{log.Id.Int64()})
 		if err != nil {
 			logger.Error(errors.Wrapf(err, "RegistrySynchronizer: unable to batch delete upkeeps, jobID: %d", rs.job.ID))
 			continue
 		}
+		logger.Debugw(fmt.Sprintf("RegistrySynchronizer: deleted %v upkeep registrations", affected), "jobID", rs.job.ID, "txHash", txHash)
 		err = broadcast.MarkConsumed()
 		logger.ErrorIf((errors.Wrapf(err, "RegistrySynchronizer: unable to mark log as consumed, jobID: %d", rs.job.ID)))
 	}
@@ -126,9 +122,6 @@ func (rs *RegistrySynchronizer) handleUpkeepRegisteredLogs(head models.Head, don
 			continue
 		}
 		if was {
-			continue
-		}
-		if !head.IsInChain(broadcast.RawLog().BlockHash) {
 			continue
 		}
 		log, ok := broadcast.DecodedLog().(*keeper_registry_wrapper.KeeperRegistryUpkeepRegistered)
@@ -167,9 +160,6 @@ func (rs *RegistrySynchronizer) handleUpkeepPerformedLogs(head models.Head, done
 			continue
 		}
 		if was {
-			continue
-		}
-		if !head.IsInChain(broadcast.RawLog().BlockHash) {
 			continue
 		}
 		log, ok := broadcast.DecodedLog().(*keeper_registry_wrapper.KeeperRegistryUpkeepPerformed)